+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Fix a crash during login when user's profile does not exist (#77)
 - Drop API support for GET, DELETE requests with a request body
 - Add missing @csrf_exempt annotations in API views
+- Fix the ping handler to reject status codes > 255
 
 ## v1.22.0 - 2020-08-06